workspace/requirements.txt
# Flask>=3.0.0,<3.1.0
# flask-cors>=4.0.0,<5.0.0
# supabase>=2.6.0,<3.0.0
# python-dotenv>=1.0.1,<2.0.0
Flask>=3.0.0,<3.1.0
flask-cors>=4.0.0,<5.0.0
mysql-connector-python>=8.2.0,<9.0.0
python-dotenv>=1.0.1,<2.0.0